basic index with CDNs
the CDNs in this will enable us to use jquery and bootstrap without having the local source files. in other words we can choose a bootstrap theme from https://startbootstrap.com/template-categories/all/ or somewhere else, and then we can branch off on any computer and the site will still work without adding the source files to the index directory.